requirements.txt
django==2.2.10
pillow==6.2.1
dj-config-url==0.1.1
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
django-absoluteuri==1.3.0
requests
